+++ unit-tests/fixtures/stubs/issues/expected/Issue_1922.zep.php
@@ -2,8 +2,7 @@
 
 namespace Stubs;
 
-
-class Test
+class Issue_1922
 {
 
     /**