+++ unit-tests/sharness/t0005-stubs.sh
@@ -6,11 +6,35 @@ test_description="Test generate IDE stubs"
 source ./setup.sh
 
 # See: https://github.com/phalcon/zephir/issues/1922
-test_expect_success "Should fail when there are no required modules" "
-  cd $FIXTURESDIR/stubs/issue1922 &&
+test_expect_success "Should properly generate type hint" "
+  cd $FIXTURESDIR/stubs/issues &&
   zephirc generate --no-ansi 2>&1 >/dev/null &&
   zephirc stubs --no-ansi 2>&1 >/dev/null &&
-  test_cmp expected/Test.zep.php ide/0.0.1/Stubs/Test.zep.php
+  test_cmp expected/Issue_1922.zep.php ide/0.0.1/Stubs/Issue_1922.zep.php
+"
+
+# See: https://github.com/phalcon/zephir/issues/1778
+test_expect_success "Should properly namespace imports (use block)" "
+  cd $FIXTURESDIR/stubs/issues &&
+  zephirc generate --no-ansi 2>&1 >/dev/null &&
+  zephirc stubs --no-ansi 2>&1 >/dev/null &&
+  test_cmp expected/Issue_1778.zep.php ide/0.0.1/Stubs/Issue_1778.zep.php
+"
+
+# See: https://github.com/phalcon/zephir/issues/1907
+test_expect_success "Should properly generate Namespace for extends" "
+  cd $FIXTURESDIR/stubs/issues &&
+  zephirc generate --no-ansi 2>&1 >/dev/null &&
+  zephirc stubs --no-ansi 2>&1 >/dev/null &&
+  test_cmp expected/Exception.zep.php ide/0.0.1/Stubs/Exception.zep.php
+"
+
+# See: https://github.com/phalcon/zephir/issues/1907
+test_expect_success "Should properly generate Namespace for extends (slash)" "
+  cd $FIXTURESDIR/stubs/issues &&
+  zephirc generate --no-ansi 2>&1 >/dev/null &&
+  zephirc stubs --no-ansi 2>&1 >/dev/null &&
+  test_cmp expected/Issue_1907.zep.php ide/0.0.1/Stubs/Issue_1907.zep.php
 "
 
 test_done